now the image background works
now the image background works

modified angular json with the suggestion

"assets": [
  "src/favicon.ico",
  "src/assets"
]

then the root itself was not working. and then updated the angular json file to original state and did a re-run.

now it is working as expected - if you would notice there is no diff in code pre and post.